+++ src/scripts/features/backgrounds/index.ts
@@ -1,17 +1,20 @@
 import { initCreditEvents, toggleCredits, updateCredits } from './credits.ts'
+import { handleBackgroundActions } from '../contextmenu.ts'
 import { applyUrls, getUrlsAsCollection, initUrlsEditor } from './urls.ts'
 import { TEXTURE_RANGES } from './textures.ts'
 import { PROVIDERS } from './providers.ts'
 import {
 	addLocalBackgrounds,
-	imageFromLocalFiles,
 	initFilesSettingsOptions,
 	lastUsedBackgroundFiles,
 	localFilesCacheControl,
+	mediaFromFiles,
+	setCurrentVideo,
 } from './local.ts'
 
 import { daylightPeriod, needsChange, userDate } from '../../shared/time.ts'
 import { colorInput, turnRefreshButton } from '../../shared/dom.ts'
+import { networkForm } from '../../shared/form.ts'
 import { rgbToHex } from '../../shared/generic.ts'
 import { debounce } from '../../utils/debounce.ts'
 import { BROWSER } from '../../defaults.ts'
@@ -20,7 +23,6 @@ import { storage } from '../../storage.ts'
 import type { Background, BackgroundImage, BackgroundVideo, Frequency } from '../../../types/shared.ts'
 import type { Backgrounds, Sync } from '../../../types/sync.ts'
 import type { Local } from '../../../types/local.ts'
-import { networkForm } from '../../shared/form.ts'
 
 type BackgroundSize = 'full' | 'medium' | 'small'
 
@@ -189,7 +191,7 @@ export async function backgroundUpdate(update: BackgroundUpdate): Promise<void> 
 		storage.local.set({ backgroundCompressFiles: update.compress })
 
 		const ids = lastUsedBackgroundFiles(local.backgroundFiles)
-		const image = await imageFromLocalFiles(ids[0], local, undefined)
+		const image = await mediaFromFiles(ids[0], local, undefined)
 
 		applyBackground(image)
 	}
@@ -614,7 +616,7 @@ export function applyBackground(media?: string | Background, res?: BackgroundSiz
 
 	const mediaWrapper = document.getElementById('background-media') as HTMLDivElement
 	let resolution = res ? res : detectBackgroundSize()
-	let item: HTMLDivElement
+	let item: HTMLElement
 
 	if (media.format === 'image') {
 		// disables blur compression for animated gifs (flawed since some gifs aren't animated)
@@ -622,10 +624,9 @@ export function applyBackground(media?: string | Background, res?: BackgroundSiz
 		const src = media.urls[resolution]
 		item = createImageItem(src, media)
 	} else {
-		const opacity = 4 //s
-		const duration = 1000 * (media.duration - opacity)
+		const fade = 4000 //ms
 		const src = media.urls[resolution]
-		item = createVideoItem(src, media, duration)
+		item = createVideoItem(src, media, fade)
 	}
 
 	item.dataset.res = resolution
@@ -666,63 +667,35 @@ function createImageItem(src: string, media: BackgroundImage, callback?: () => v
 
 	div.style.backgroundImage = `url(${src})`
 
-	if (media.size) {
-		div.style.backgroundSize = media.size
-	}
-	if (media.x) {
-		div.style.backgroundPositionX = media.x
-	}
-	if (media.y) {
-		div.style.backgroundPositionY = media.y
+	if (media?.file?.position) {
+		const { size, x, y } = media.file.position
+
+		div.style.backgroundSize = size
+		div.style.backgroundPositionX = x
+		div.style.backgroundPositionY = y
 	}
 
 	return div
 }
 
-function createVideoItem(src: string, media: BackgroundVideo, duration: number): HTMLDivElement {
+function createVideoItem(src: string, media: BackgroundVideo, duration: number): HTMLElement {
 	const backgroundsWrapper = document.getElementById('background-wrapper')
-	const div = document.createElement('div')
-	let videoInterval = 0
+	const looper = setCurrentVideo(src, duration, 1)
+	const container = looper.getContainer()
 
-	const prependVideo = () => {
-		return new Promise((resolve) => {
-			const vid = document.createElement('video')
+	looper.loop()
 
-			vid.addEventListener('canplay', () => {
-				backgroundsWrapper?.classList.remove('hidden')
-				updateCredits(media)
-				resolve(true)
-			})
+	if (media?.file?.video) {
+		const { playbackRate, zoom, fade } = media.file.video
 
-			vid.src = src
-			vid.volume = 0
-			vid.muted = true
-			vid.autoplay = true
-			vid.playbackRate = 1
-			div.prepend(vid)
-		})
+		container.style.transform = `scale(${zoom})`
+		looper.setPlaybackRate(playbackRate)
+		looper.setFadeTime(fade)
 	}
 
-	const removeVideo = () => {
-		div?.lastElementChild?.classList.add('hiding')
-		setTimeout(() => div?.lastElementChild?.remove(), 4000)
-	}
+	backgroundsWrapper?.classList.remove('hidden')
 
-	const loopVideo = async () => {
-		if (document.body.contains(div)) {
-			await prependVideo()
-			removeVideo()
-			return
-		}
-
-		clearInterval(videoInterval)
-	}
-
-	prependVideo().then(() => {
-		videoInterval = setInterval(loopVideo, duration)
-	})
-
-	return div
+	return container
 }
 
 function preloadBackground(media: Background | undefined, res?: BackgroundSize) {
@@ -933,18 +906,10 @@ function createProviderSelect(backgrounds: Backgrounds) {
 	}
 }
 
-function handleBackgroundActions(backgrounds: Backgrounds) {
-	const type = backgrounds.type
-	const freq = backgrounds.frequency
-	document.getElementById('background-actions')?.classList.toggle('shown', type !== 'color')
-	document.getElementById('b_interface-background-pause')?.classList.toggle('paused', freq === 'pause')
-	document.getElementById('b_interface-background-download')?.classList.toggle('shown', type === 'images')
-}
-
 async function blurResolutionControl(sync: Sync, local: Local) {
 	if (sync.backgrounds.type === 'files') {
 		const ids = lastUsedBackgroundFiles(local.backgroundFiles)
-		const image = await imageFromLocalFiles(ids[0], local)
+		const image = await mediaFromFiles(ids[0], local)
 		applyBackground(image, 'full')
 		return
 	}
@@ -968,8 +933,8 @@ async function blurResolutionControl(sync: Sync, local: Local) {
 async function getCurrentBackgrounds(sync: Sync, local: Local) {
 	if (sync.backgrounds.type === 'files') {
 		const ids = lastUsedBackgroundFiles(local.backgroundFiles)
-		const current = await imageFromLocalFiles(ids[0], local)
-		const next = await imageFromLocalFiles(ids[1], local)
+		const current = await mediaFromFiles(ids[0], local)
+		const next = await mediaFromFiles(ids[1], local)
 		return [current, next]
 	}
 	if (sync.backgrounds.type === 'images') {