+++ src/scripts/features/links/index.ts
@@ -1,6 +1,7 @@
 import { addGroup, changeGroupTitle, deleteGroup, initGroups, moveGroups, toggleGroups } from './groups.ts'
 import { initBookmarkSync, syncBookmarks } from './bookmarks.ts'
-import { openEditDialog } from './edit.ts'
+import { openContextMenu } from '../contextmenu.ts'
+import { storeIconFile } from './fileicons.ts'
 import { folderClick } from './folders.ts'
 import { startDrag } from './drag.ts'
 import {
@@ -20,7 +21,8 @@ import { eventDebounce } from '../../utils/debounce.ts'
 import { tradThis } from '../../utils/translations.ts'
 import { storage } from '../../storage.ts'
 
-import type { Link, LinkElem, LinkFolder } from '../../../types/shared.ts'
+import type { Link, LinkElem, LinkFolder, LinkIcon } from '../../../types/shared.ts'
+import type { Local } from '../../../types/local.ts'
 import type { Sync } from '../../../types/sync.ts'
 
 type AddLinks = {
@@ -32,8 +34,9 @@ type AddLinks = {
 type UpdateLink = {
 	id: string
 	url?: string
-	icon?: string
 	title: string
+	icon?: LinkIcon
+	file?: File
 }
 
 type AddGroups = {
@@ -94,11 +97,16 @@ type LinkGroups = {
 	div: HTMLDivElement | null
 }[]
 
+type LinksInit = {
+	sync: Sync
+	local: Local
+}
+
 const domlinkblocks = document.getElementById('linkblocks') as HTMLDivElement
 let initIconList: [HTMLImageElement, string][] = []
 let selectallTimer = 0
 
-export async function quickLinks(init?: Sync, event?: LinksUpdate) {
+export async function quickLinks(init?: LinksInit, event?: LinksUpdate) {
 	if (event) {
 		linksUpdate(event)
 		return
@@ -108,33 +116,35 @@ export async function quickLinks(init?: Sync, event?: LinksUpdate) {
 		return
 	}
 
-	// set class before appendBlock, cannot be moved
-	domlinkblocks.classList.add(init.linkstyle ?? 'large')
-	domlinkblocks.classList.toggle('titles', init.linktitles)
-	domlinkblocks.classList.toggle('backgrounds', init.linkbackgrounds)
-	domlinkblocks.classList.toggle('hidden', !init.quicklinks)
+	const { sync, local } = init
 
-	if (init.linkgroups.synced.length > 0) {
-		await initBookmarkSync(init)
+	// set class before appendBlock, cannot be moved
+	domlinkblocks.classList.add(sync.linkstyle ?? 'large')
+	domlinkblocks.classList.toggle('titles', sync.linktitles)
+	domlinkblocks.classList.toggle('backgrounds', sync.linkbackgrounds)
+	domlinkblocks.classList.toggle('hidden', !sync.quicklinks)
+
+	if (sync.linkgroups.synced.length > 0) {
+		await initBookmarkSync(sync)
 	}
 
-	initGroups(init, !!init)
-	initRows(init.linksrow, init.linkstyle)
-	initblocks(init, true)
+	initGroups(sync, !!init)
+	initRows(sync.linksrow, sync.linkstyle)
+	initblocks(sync, local)
 }
 
 // Initialisation
 
-export function initblocks(data: Sync, isInit?: true): true {
-	const allLinks = Object.values(data).filter((val) => isLink(val)) as Link[]
-	const { pinned, synced, selected } = data.linkgroups
+export function initblocks(sync: Sync, local?: Local): true {
+	const allLinks = Object.values(sync).filter((val) => isLink(val)) as Link[]
+	const { pinned, synced, selected } = sync.linkgroups
 	const activeGroups: LinkGroups = []
 
 	for (const group of [...pinned, selected]) {
 		const div = document.querySelector<HTMLDivElement>(`.link-group[data-group="${group}"]`)
 		const folder = div?.dataset.folder
 		const lis: HTMLLIElement[] = []
-		const links = folder ? getLinksInFolder(data, folder) : getLinksInGroup(data, group)
+		const links = folder ? getLinksInFolder(sync, folder) : getLinksInGroup(sync, group)
 
 		activeGroups.push({
 			lis,
@@ -190,11 +200,12 @@ export function initblocks(data: Sync, isInit?: true): true {
 			}
 
 			li = isElem(link)
-				? createElem(link, data.linknewtab, data.linkstyle)
-				: createFolder(link, linksInFolders, data.linkstyle)
+				? createElem(link, sync.linknewtab, sync.linkstyle)
+				: createFolder(link, linksInFolders, sync.linkstyle)
 
 			fragment.appendChild(li)
-			li.addEventListener('keyup', openEditDialog)
+
+			li.addEventListener('keyup', openContextMenu)
 
 			if (!group.synced) {
 				li.addEventListener('click', selectAll)
@@ -204,7 +215,7 @@ export function initblocks(data: Sync, isInit?: true): true {
 		}
 
 		if (folderid) {
-			linktitle.textContent = (data[folderid] as LinkFolder).title
+			linktitle.textContent = (sync[folderid] as LinkFolder).title
 		} else {
 			linktitle.textContent = group.title
 		}
@@ -226,7 +237,14 @@ export function initblocks(data: Sync, isInit?: true): true {
 		}
 	}
 
-	createIcons(isInit)
+	if (local) {
+		createIcons(local)
+	} else {
+		storage.local.get().then((local) => {
+			createIcons(local)
+		})
+	}
+
 	displayInterface('links')
 
 	return true
@@ -248,6 +266,7 @@ function createFolder(link: LinkFolder, folderChildren: Link[], style: Sync['lin
 	li.id = link._id
 	span.textContent = createTitle(link)
 	li.addEventListener('mouseup', folderClick)
+	li.addEventListener('keydown', folderClick)
 
 	for (let i = 0; i < linksInThisFolder.length; i++) {
 		const img = imgs[i]
@@ -255,7 +274,7 @@ function createFolder(link: LinkFolder, folderChildren: Link[], style: Sync['lin
 		const isIconShown = img && isElem(elem) && style !== 'text'
 
 		if (isIconShown) {
-			initIconList.push([img, elem.icon ?? getDefaultIcon(elem.url)])
+			initIconList.push([img, getIconFromLinkElem(elem)])
 		}
 	}
 
@@ -277,12 +296,7 @@ function createElem(link: LinkElem, openInNewtab: boolean, style: Sync['linkstyl
 	span.textContent = createTitle(link)
 
 	if (style !== 'text') {
-		const iconurl = link.icon ?? ''
-		const refresh = new Date(Number.parseInt(iconurl))?.getTime()
-		const isDefaultRefreshed = Number.isInteger(refresh)
-		const icon = !isDefaultRefreshed && link.icon ? link.icon : getDefaultIcon(link.url, refresh)
-
-		initIconList.push([img, icon])
+		initIconList.push([img, getIconFromLinkElem(link)])
 	}
 
 	if (openInNewtab) {
@@ -292,16 +306,23 @@ function createElem(link: LinkElem, openInNewtab: boolean, style: Sync['linkstyl
 	return li
 }
 
-function createIcons(isInit?: true) {
-	const loadingTimeout = isInit ? 400 : 0
-
+function createIcons(local: Local) {
 	for (const [img, url] of initIconList) {
-		img.src = url
+		if (url.startsWith('link')) {
+			img.src = local[`x-icon-${url}`] ?? ''
+		} else {
+			img.src = url
+		}
 	}
 
 	setTimeout(() => {
-		const incomplete = initIconList.filter(([img]) => !img.complete)
+		// naturalWidth is needed here because complete doesn't tell the whole story
+		// it only says if it's finished loading or not, even an error code will say "complete"
+		const incomplete = initIconList.filter(
+			([img]) => !img.complete || img.naturalWidth === 0,
+		)
 
+		// if images still haven't loaded after 400ms
 		for (const [img, url] of incomplete) {
 			img.src = 'src/assets/interface/loading.svg'
 
@@ -311,11 +332,24 @@ function createIcons(isInit?: true) {
 				img.src = url
 			})
 
+			// if obvious error (dead link...), shows fallback
+			newimg.addEventListener('error', () => {
+				img.src = 'https://services.bonjourr.fr/favicon/blob/error'
+			})
+
 			newimg.src = url
+
+			// If image still isn't responding after 5s, gives up
+			setTimeout(() => {
+				if (!newimg.complete && newimg.naturalWidth === 0) {
+					console.error('Icon link took too long to load: ' + url)
+					img.src = 'https://services.bonjourr.fr/favicon/blob/error'
+				}
+			}, 5000)
 		}
 
 		initIconList = []
-	}, loadingTimeout)
+	}, 400)
 }
 
 function initRows(row: number, style: string) {
@@ -451,7 +485,12 @@ function linkSubmission(args: SubmitLink | SubmitFolder, data: Sync): Sync {
 
 	if (type === 'link') {
 		for (const link of args.links) {
-			newlinks.push(validateLink(link.title, link.url, link.group))
+			newlinks.push(validateLink(
+				link.title,
+				link.url,
+				// if no group is specified, adds to the selected one
+				link.group || data.linkgroups.selected,
+			))
 		}
 	}
 
@@ -484,11 +523,13 @@ function linkSubmission(args: SubmitLink | SubmitFolder, data: Sync): Sync {
 		data[link._id] = link
 	}
 
-	const correctdata = correctLinksOrder(data)
+	const newsync = correctLinksOrder(data)
 
-	initblocks(correctdata)
+	storage.local.get().then((local) => {
+		initblocks(newsync, local)
+	})
 
-	return correctdata
+	return newsync
 }
 
 function addLinkFolder(ids: string[], title?: string, group?: string): LinkFolder[] {
@@ -521,10 +562,11 @@ function addLinkFolder(ids: string[], title?: string, group?: string): LinkFolde
 	]
 }
 
-function updateLink({ id, title, icon, url }: UpdateLink, data: Sync): Sync {
+function updateLink({ id, title, icon, url, file }: UpdateLink, data: Sync): Sync {
 	const titledom = document.querySelector<HTMLSpanElement>(`#${id} span`)
 	const icondom = document.querySelector<HTMLImageElement>(`#${id} img`)
 	const urldom = document.querySelector<HTMLAnchorElement>(`#${id} a`)
+
 	const link = data[id] as Link
 
 	if (titledom && title !== undefined) {
@@ -532,22 +574,56 @@ function updateLink({ id, title, icon, url }: UpdateLink, data: Sync): Sync {
 		titledom.textContent = link.title
 	}
 
-	if (!link.folder) {
-		if (icondom) {
-			const url = (icon ? stringMaxSize(icon, 7500) : undefined) ?? getDefaultIcon(link.url)
+	if (isElem(link)) {
+		if (icondom && icon) {
 			const img = document.createElement('img')
-
-			link.icon = url ? url : undefined
+			const currentSrc = icondom.src
+			let url = getDefaultIcon(link.url)
 
 			icondom.src = 'src/assets/interface/loading.svg'
 
 			img.onload = () => {
-				if (icondom) {
-					icondom.src = url
+				icondom.src = img.src
+			}
+
+			if (icon.type === 'auto') {
+				icon.value = undefined
+				img.src = url
+			}
+
+			if (icon.type === 'url') {
+				if (icon.value && stringMaxSize(icon.value, 7500)) {
+					url = icon.value
+					img.src = url
+				} else {
+					console.error(`There was a problem with this icon URL: ${icon.value}`)
 				}
 			}
 
-			img.src = url
+			if (icon.type === 'file') {
+				const currentIcon = link.icon
+				const noNewOrCurrentFile = !file && !currentIcon?.value
+				const noNewButHasCurrentFile = !file && (currentIcon?.type === 'file') && !!currentIcon?.value
+
+				if (noNewOrCurrentFile) {
+					throw new Error('Chose file but no file uploaded')
+				}
+
+				if (noNewButHasCurrentFile) {
+					icon = currentIcon
+					img.src = currentSrc
+				}
+
+				if (file) {
+					url = id
+
+					storeIconFile(id, file).then((uri) => {
+						img.src = uri
+					})
+				}
+			}
+
+			link.icon = icon
 		}
 
 		if (titledom && urldom && url !== undefined) {
@@ -558,6 +634,7 @@ function updateLink({ id, title, icon, url }: UpdateLink, data: Sync): Sync {
 	}
 
 	data[id] = link
+
 	return data
 }
 
@@ -631,6 +708,12 @@ function deleteLinks(ids: string[], data: Sync): Sync {
 			}
 		}
 
+		if (isElem(link)) {
+			if (link.icon?.type === 'file') {
+				storage.local.remove(`x-icon-${id}`)
+			}
+		}
+
 		delete data[id]
 	}
 
@@ -656,6 +739,7 @@ function moveToGroup({ ids, target }: MoveToGroup, data: Sync): Sync {
 	}
 
 	const correctdata = correctLinksOrder(data)
+
 	initblocks(correctdata)
 	return correctdata
 }
@@ -665,7 +749,15 @@ function refreshIcons(ids: string[], data: Sync): Sync {
 		const link = data[id] as LinkElem
 
 		if (link._id) {
-			link.icon = Date.now().toString()
+			const unixDate = Date.now().toString()
+
+			if (!link.icon || link.icon.type === 'auto') {
+				link.icon = link.icon ?? { type: 'auto', value: '' } // when link was just added, it doesn't have the icon property, so creates it
+				link.icon.value = getDefaultIcon(link.url) + `?r=${unixDate}`
+			} else if (link.icon.type === 'url') {
+				link.icon.value = `${link.icon.value}?r=${unixDate}`
+			}
+
 			data[id] = link
 		}
 	}
@@ -791,6 +883,17 @@ function correctLinksOrder(data: Sync): Sync {
 	}
 
 	return data
+}
+
+function getIconFromLinkElem(link: LinkElem): string {
+	if (!link.icon?.value) {
+		return getDefaultIcon(link.url)
+	}
+	if (link.icon.type === 'file') {
+		return link._id
+	}
+
+	return link.icon.value
 }
 
 function isLinkStyle(s: string): s is Sync['linkstyle'] {